requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
fakeredis==2.34.1
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.26.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
python-dotenv==1.2.2
python-multipart==0.0.22
qrcode==8.2
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sse-starlette==3.3.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
